More bug fixes (#24)
* Some Facts converted to Theory

* - Added tag Delete
- Added new tests
- Added template validations
- Improved based test class
- Fixed bugs with grouping for nested groups
- Fixed issues with Protect tag
- Fixed issues with MergeLabels option for Group tag

* Cosmetic changes

* ColsFit tag fix

* changes for review requesting

* Added test ParseExceptionMessageShouldBeUnknownIdentifier

* Fix syntax error

+++ tests/ClosedXML.Report.Tests/TestModels/DbDemos.cs
@@ -63,7 +63,7 @@ namespace ClosedXML.Report.Tests.TestModels
     public partial class customer
     {
         [PrimaryKey, Identity]
-        public double CustNo { get; set; } // Double
+        public int CustNo { get; set; } // int
         [Column, Nullable]
         public string Company { get; set; } // text(30)
         [Column, Nullable]
